pyproject.toml
[project]
name = "aidebate-prototype"
version = "0.1.0"
description = "Simple LiveKit video chat prototype"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.5",
    "fastapi>=0.116.1",
    "livekit-api>=1.1.0",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
]
